requirements.txt
Flask
Markdown
requests
beautifulsoup4
pandas
python-dotenv
